backend/requirements.txt
azure-storage-blob==12.9.0
azure-identity==1.15.0
python-dotenv==1.0.1
configparser==6.0.0
requests==2.31.0
openpyxl==3.1.2
pyodbc==5.0.1
pandas==2.2.0
numpy==1.26.3
lxml==5.1.0
xlrd==2.0.1
flask==3.0.2
flask_cors==3.0.10
Flask-SQLAlchemy==3.1.1